feat: allow including chilren in sink [ENG-5531] (#6)
[ENG-5531](https://stacklet.atlassian.net/browse/ENG-5531)

### what

If you setup a sink without including the children it will only give
logs for resources directly in the root of the folder and not any
subfolders.

I had to structure it in a more complicated way because of this bug:
- https://github.com/hashicorp/terraform-provider-google/issues/10811

### why

Sometimes customers want a simple setup with all resources within a
folder to feed the log.

### testing

I setup a relay with this code:
- https://github.com/stacklet/sontek/pull/8/files

and proved it work by creating compute instances and seeing it trigger
on the AWS side:

<img width="718" height="767" alt="Screen Shot 2025-10-01 at 4 52 34 PM"
src="https://github.com/user-attachments/assets/f0c9689b-1b56-458e-93bf-55c669437e2c"
/>

with this PR the TF plan was:

```
# module.stacklet_relay.google_logging_folder_sink.folder_feed[0] will be updated in-place
  ~ resource "google_logging_folder_sink" "folder_feed" {
        id                 = "folders/492199171560/sinks/folder-audit-feed-492199171560"
      ~ include_children   = false -> true
        name               = "folder-audit-feed-492199171560"
        # (7 unchanged attributes hidden)
    }
```

### docs

Docs are in the PR


[ENG-5531]:
https://stacklet.atlassian.net/browse/ENG-5531?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: John Anderson <john@stacklet.io>